I hate memory management in garbage collected languages.
The `replace_contents_async` function doesn't copy the `content` variable.
This variable then gets garbage disposed before the function accesses it, resulting in undefined behavior.
In practice, this wrote random data to the file, which typically contained shell command errors (idk why).

The fix is to create a module scoped table that I can store the values into, which prevents them from being garbage collected.
The content is used as the key, as it should (hopefully) be unique. Perhaps this should be changed to reduce collisions.
As the content is therefore *not* garbage collected, it can then be used to index into the array and unset the value.