add error to source_path if used in a TCO manner. This was the cause of many bugs/headaches
TCO changes all the info and makes source info for the desired level inaccessible. This change bans using TCO with source_path, allowing the filename to be properly found.

Note: this could have bad effects if used in dynamic ways, making it difficult to find the cause of the error. However, this error indicates a broken expectaion and thusly a bug.